Development

+++ StudeasySystem/ejbModule/de/studeasy/schedulemanager/StudeasyScheduleService.java
@@ -100,29 +100,51 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	
 	
 	//---------------------------CREATE HOMEWORK------------------------------------------------------------
-	//TODO Muss noch richtig implementiert werden.
+	/**
+	 * Legt eine neue Hausaufgabe für eine bestimmtest Fach an.
+	 * Mit der sessionID wird auf Berechtigung geprüft.
+	 * Wenn der User kein Lehrer ist, wird keine Hausaufgabe erzeugt und im BooleanResponse der Error-Code 30 gesetzt.
+	 */
 	@Override
 	public BooleanResponse createHomework(int sessionID, int lessonID, String description)  {
 		
 		BooleanResponse response = new BooleanResponse();
+		IPerson user = dao.findPersonByID(dao.findSessionByID(sessionID).getUserID());
 		
-	//TODO mit sessionID noch auf Berechtigung prüfen
+		if(user instanceof ITeacher) {
 			ILesson lesson = dao.findLessonByID(lessonID);
 			IHomework homework = new Homework(description, lesson);
 			lesson.addNewHomework(homework);
 			response.setSuccessfull(true);
+		}
+		else {
+			response.setSuccessfull(false);
+			response.setMessage("Der angemeldete User ist keine Lehrer und darf somit keine Hausaufgaben anlegen.");
+			response.setReturnCode(30);
+		}
 		return response;
 	}
 	//---------------------------REMOVE HOMEWORK------------------------------------------------------------
+	/**
+	 * Löscht eine Hausaufgabe.
+	 * Mit der sessionID wird auf Berechtigung geprüft.
+	 * Wenn der User kein Lehrer ist, wird die Hausaufgabe nicht gelöscht und im BooleanResponse der Error-Code 30 gesetzt.
+	 */
 	@Override
 	public BooleanResponse removeHomework(int sessionID, int homeworkID)  {
+		BooleanResponse response = new BooleanResponse();
+		boolean successfull = false;
+		IPerson user = dao.findPersonByID(dao.findSessionByID(sessionID).getUserID());
 		
-			boolean successfull = dao.removeHomeworkByID(homeworkID);
+		if(user instanceof ITeacher) {
+			successfull = dao.removeHomeworkByID(homeworkID);
+		}
+		else {
+			response.setMessage("Der angemeldete User ist keine Lehrer und darf somit keine Hausaufgaben anlegen.");
+			response.setReturnCode(30);
+		}
 			
-			BooleanResponse response = new BooleanResponse();
-			
-			response.setSuccessfull(successfull);
-		
+		response.setSuccessfull(successfull);
 		return response;
 	}
 	
@@ -131,7 +153,6 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	 * Gibt eine Liste der Unterrichtsstunden an einem bestimmten Tag
 	 * für eine bestimmte Person zurück.
 	 * Die Liste ist leer, wenn an dem Tag kein für die Person kein Unterricht ist.
-	 * Wenn null zurückgeben wird, waren die Parameter falsch.
 	 */
 	
 	@Override
@@ -145,6 +166,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 		try {
 			StudeasySession session = getSession(sessionID);
 			IPerson person = dao.findPersonByID(session.getUserID());
+			
 			if(person instanceof IPupil) {
 				IPupil pupil = (IPupil) person;
 				lessons = pupil.getCourse().getLessons();	
@@ -193,8 +215,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	/** 
 	 * Gibt eine Liste der Unterrichtsstunden in einem bestimmten Zeitraum
 	 * für eine bestimmte Schulklasse in einem bestimmten Fach zurück.
-	 * Die Liste ist leer, wenn in dem Zeitruam kein für die Schulklasse in dem Fach kein Unterricht ist.
-	 * Wenn null zurückgeben wird, waren die Parameter falsch.
+	 * Die Liste ist leer, wenn in dem Zeitraum für die Schulklasse in dem Fach kein Unterricht ist.
 	 * Das startDate und endDate sind inklusive. Also Unterrichtsstunden die am 
 	 * start- oder endDate stattfinden befinden sich in der zurückgegebenen Liste.
 	 */
@@ -241,7 +262,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	/**
 	 * Gibt eine Liste von Hausaufgaben für einen bestimmten Schüler zu einem bestimmten Zeitraum zurück.
 	 * Die Liste ist leer, wenn dieser Schüler zu diesem Zeitraum keine Hausaufgaben hat.
-	 * Es wird null zurückgegeben, wenn die personID nicht zu einem Schüler gehört.
+	 * Wenn der angemeldete User kein Schüler ist wird im HomeworkListResponse der Error-Code auf 30 gesetzt.
 	 */
 	@Override
 	public HomeworkListResponse getHomeworksForPupil(int sessionID, Date startDate,
@@ -269,13 +290,15 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 				
 				response.setHomeworkList(dto.makeHomeworkDTO(homeworks));
 				}
-			else
-				response.setHomeworkList( null);
-			
+			else {
+				response.setHomeworkList(null);
+				response.setMessage("Der angemeldete User ist keine Schüler.");
+				response.setReturnCode(30);
+			}
 			
 		}
 		catch (StudeasyException e) {
-			response.setHomeworkList( null);
+			response.setHomeworkList(null);
 			response.setReturnCode(e.getErrorCode());
 			response.setMessage(e.getMessage());
 		}
